pyproject.toml
[tool.poetry]
name = "session_token_auth_demo"
version = "0.1.0"
description = ""
authors = ["Josef Rousek <josef.rousek@gmail.com>"]

[tool.poetry.dependencies]
python = "3.8.6"
Django = "^3.1.5"
whitenoise = "^5.2.0"
waitress = "^1.4.4"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
python-dotenv = {extras = ["cli"], version = "^0.15.0"}
django-shopify-auth = { git = "https://github.com/discolabs/django-shopify-auth.git", branch = "feature/cookieless-auth" }

[tool.poetry.dev-dependencies]
pre-commit = "^2.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
